+++ include/pokemon_icon.h
@@ -3,9 +3,9 @@
 
 extern const struct SpritePalette gMonIconPaletteTable[];
 
-const u8 *GetMonIconTiles(u16 species, u32 personality);
+const u32 *GetMonIconTiles(u16 species, u32 personality);
 void TryLoadAllMonIconPalettesAtOffset(u16 offset);
-const u8 *GetMonIconPtr(u16 speciesId, u32 personality);
+const u32 *GetMonIconPtr(u16 speciesId, u32 personality);
 u16 GetIconSpecies(u16 species, u32 personality);
 u16 GetUnownLetterByPersonality(u32 personality);
 u16 GetIconSpeciesNoPersonality(u16 speciesId);
